Add prefetch partitions functional test
Functional tests to check api compatibility.

Relates-To: OLPEDGE-2343
